projects/multi-agent-career-advisor/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
av==16.0.1
azure-ai-agents==1.2.0b6
azure-ai-projects==1.0.0
azure-core==1.37.0
azure-cosmos==4.5.1
azure-identity==1.25.1
azure-storage-blob==12.27.1
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudevents==1.12.0
cryptography==46.0.3
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
MarkupSafe==3.0.3
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==2.4.0
openai==1.109.1
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
parse==1.20.2
pathable==0.4.4
pillow==12.0.0
plotly==6.5.0
prance==25.4.8.0
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pybars4==0.9.13
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
PyGithub==2.8.1
PyJWT==2.10.1
pylibsrtp==1.0.0
PyMeta3==0.5.1
PyNaCl==1.6.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.1
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rpds-py==0.30.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
schedule==1.2.2
scipy==1.16.3
semantic-kernel==1.39.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.22.0
zipp==3.23.0
